Fix bug caused by double-escaped parenthesis
The prompt showing the available mounting points is empty.
Firstly the '-not' options used in find must be seperated by a space,
but that alone doesn't seem to fix the issue. Turning on bash's
debugging mode reveals this:

find /mnt -maxdepth 3 -type d -empty -not '\(' ...

The full line is not shown due it's length, but regardless the issue is
due to the '\(' term. The single-quotes shouldn't be here and this
causes find to fail. Escaping the parenthesis only once results in '('
and ')' which is fine.